pyproject.toml
[project]
name = "web-aware-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.119.0",
    "google-genai>=1.45.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pgvector>=0.4.1",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "redis>=6.4.0",
    "sqlalchemy>=2.0",
    "uvicorn>=0.37.0",
    "voyageai>=0.3.5",
]
